Add a patch to fix logrotate with ganeti
The problem is that `/etc/logrotate.d/ganeti` calls `/usr/lib64/ganeti/daemon-util rotate-all-logs` which tries to stop all daemons including the on-demand `ganeti-metad` daemon, and it fails because the `/run/ganeti/ganeti-metad.pid` doesn't exists, and the whole script exits with nonzero exit status because of `set -e`.

The fix is to to change `list_stop_daemons` to only list actually running daemons.